refactor(quantum_rng_v3): split 1176-line monolith into cohesive modules
Replace src/quantum_rng/quantum_rng_v3.c with a set of single-responsibility
translation units, per soul.md §10 (no monoliths, self-contained modules):

  v3_internal.h     - private shared static declarations + common includes
  v3_config.c       - default configuration
  v3_init.c         - entropy callback + init/free lifecycle
  v3_extract.c      - quantum-state evolution + direct/grover entropy extraction
  v3_generation.c   - bytes/uint64/double/range output generation
  v3_grover.c       - the 3 Grover sampling APIs
  v3_verify.c       - Bell-CHSH verification + entanglement entropy
  v3_mode_stats.c   - mode control + stats retrieval/print
  v3_utils.c        - version/error/mode strings + seed-based init

Verified: libs build clean (-Wall -Wextra, zero warnings) and qrng_v3_test
passes every suite (identical CHSH=2.84). Pure refactor, no behavior change.

The repo Makefile uses $(wildcard src/quantum_rng/*.c), so the new modules
are auto-compiled and linked with no Makefile edit required.
