--- src/quantum_rng/quantum_rng_v3.c
@@ -1,1176 +1,0 @@
-#include "quantum_rng_v3.h"
-#include "quantum_constants.h"
-#include "simd_ops.h"
-#include "../common/secure_memory.h"
-#include "../common/validation.h"
-#include <stdlib.h>
-#include <string.h>
-#include <stdio.h>
-#include <math.h>
-#include <time.h>
-
-/**
- * @file quantum_rng_v3.c
- * @brief Unified Quantum RNG v3.0 - Production implementation
- * 
- * CRITICAL ARCHITECTURE:
- * This resolves the entropy circular dependency through layered design:
- * 
- * Layer 1 (Base): Hardware entropy pool
- *   - RDSEED, RDRAND, /dev/random, jitter
- *   - Health tested (NIST SP 800-90B)
- *   - Background generation thread
- * 
- * Layer 2 (Quantum): Quantum simulation
- *   - Uses Layer 1 for measurement sampling
- *   - No circular dependency!
- *   - Bell-verified quantum evolution
- * 
- * Layer 3 (Output): Conditioned random output
- *   - Quantum-evolved entropy
- *   - Final mixing and conditioning
- *   - Cryptographic quality
- */
-
-// ============================================================================
-// CONFIGURATION
-// ============================================================================
-
-void qrng_v3_get_default_config(qrng_v3_config_t *config) {
-    if (!config) return;
-    
-    // Quantum simulation defaults
-    config->num_qubits = 8;  // 256-dimensional state space
-    config->mode = QRNG_V3_MODE_DIRECT;
-    
-    // Bell test monitoring (disable for benchmarks, enable for production)
-    config->enable_bell_monitoring = 0;  // Disabled by default for performance
-    config->bell_test_interval = 1024 * 1024;  // Test every 1MB when enabled
-    // Alert threshold for the monitored CHSH value. A healthy source produces
-    // ~2.83 (Tsirelson) and the classical bound is 2.0, so 2.1 flags a genuine
-    // collapse toward classical while leaving ample margin for the sampling
-    // noise of a finite-measurement CHSH estimate (avoids spurious failures).
-    config->min_acceptable_chsh = 2.1;
-    
-    // Grover optimization
-    config->enable_grover_cache = 1;
-    config->grover_cache_size = 256;
-    
-    // Entropy pool
-    config->entropy_pool_size = 64 * 1024;  // 64KB pool
-    config->enable_background_entropy = 1;
-    
-    // Output buffer (larger = fewer refills = better performance)
-    config->output_buffer_size = 64 * 1024;  // 64KB buffer
-    
-    // Performance
-    config->enable_simd = 1;
-    config->enable_performance_monitoring = 0;  // Disabled by default (overhead)
-}
-
-// ============================================================================
-// ENTROPY CALLBACK (Resolves circular dependency!)
-// ============================================================================
-
-/**
- * @brief Entropy callback for quantum measurements
- * 
- * This function provides hardware entropy to the quantum simulation layer.
- * Key: Uses entropy_pool (Layer 1) for quantum measurements (Layer 2).
- * No circular dependency because entropy_pool doesn't use quantum simulation!
- */
-static int entropy_pool_callback(void *user_data, uint8_t *buffer, size_t size) {
-    entropy_pool_ctx_t *pool = (entropy_pool_ctx_t *)user_data;
-    if (!pool) return -1;
-    
-    return entropy_pool_get_bytes(pool, buffer, size);
-}
-
-// ============================================================================
-// QUANTUM STATE EVOLUTION
-// ============================================================================
-
-/**
- * @brief Evolve quantum state through random circuit
- * 
- * Applies random quantum gates to create complex quantum evolution.
- * Uses hardware entropy to select gates (no circular dependency).
- */
-static qs_error_t evolve_quantum_state(
-    quantum_state_t *state,
-    quantum_entropy_ctx_t *entropy_ctx,
-    size_t num_gates
-) {
-    if (!state || !entropy_ctx) return QS_ERROR_INVALID_STATE;
-    
-    for (size_t i = 0; i < num_gates; i++) {
-        // Get random gate selection using hardware entropy
-        uint64_t gate_selector;
-        if (quantum_entropy_get_uint64(entropy_ctx, &gate_selector) != 0) {
-            return QS_ERROR_INVALID_STATE;
-        }
-        
-        int qubit1 = gate_selector % state->num_qubits;
-        int qubit2 = (gate_selector >> 8) % state->num_qubits;
-        
-        // Select random gate type
-        int gate_type = (gate_selector >> 16) % 8;
-        
-        switch (gate_type) {
-            case 0:
-                gate_hadamard(state, qubit1);
-                break;
-            case 1:
-                if (qubit1 != qubit2) {
-                    gate_cnot(state, qubit1, qubit2);
-                }
-                break;
-            case 2: {
-                double angle = ((gate_selector >> 24) & 0xFF) * QC_2PI / 256.0;
-                gate_ry(state, qubit1, angle);
-                break;
-            }
-            case 3:
-                gate_pauli_x(state, qubit1);
-                break;
-            case 4:
-                gate_pauli_z(state, qubit1);
-                break;
-            case 5:
-                gate_s(state, qubit1);
-                break;
-            case 6:
-                if (qubit1 != qubit2) {
-                    gate_cz(state, qubit1, qubit2);
-                }
-                break;
-            case 7: {
-                double angle = ((gate_selector >> 32) & 0xFF) * QC_2PI / 256.0;
-                gate_rz(state, qubit1, angle);
-                break;
-            }
-        }
-    }
-    
-    return QS_SUCCESS;
-}
-
-/**
- * @brief Extract randomness from quantum measurements (OPTIMIZED)
- *
- * PERFORMANCE CRITICAL OPTIMIZATION:
- * Instead of evolving for every 8 bytes, we:
- * 1. Evolve state ONCE with deep circuit
- * 2. Extract MULTIPLE measurements without re-evolving
- * 3. Only re-evolve when entropy pool depleted
- *
- * This gives 10-50x speedup while maintaining quantum properties!
- */
-static int extract_quantum_entropy(
-    qrng_v3_ctx_t *ctx,
-    uint8_t *buffer,
-    size_t size
-) {
-    if (!ctx || !buffer || size == 0) return -1;
-    
-    size_t bytes_generated = 0;
-    
-    // GUARANTEED UNIQUENESS: Always evolve between measurements
-    // Balance: Batch for performance, but ensure diversity
-    const size_t measurements_per_full_evolution = 64;  // Smaller batch = more diversity
-    size_t measurements_since_evolution = measurements_per_full_evolution;  // Force first evolution
-    
-    while (bytes_generated < size) {
-        // Full evolution circuit periodically
-        if (measurements_since_evolution >= measurements_per_full_evolution) {
-            evolve_quantum_state(
-                ctx->quantum_state,
-                &ctx->entropy_ctx,
-                ctx->config.num_qubits * 2  // 2 gates per qubit
-            );
-            measurements_since_evolution = 0;
-        }
-        
-        // CRITICAL: Apply 4 gates between EVERY measurement for guaranteed uniqueness
-        // More gates = more entropy mixing = better diversity
-        uint64_t gate_selector;
-        if (quantum_entropy_get_uint64(&ctx->entropy_ctx, &gate_selector) == 0) {
-            int qubit1 = gate_selector % ctx->quantum_state->num_qubits;
-            int qubit2 = (gate_selector >> 8) % ctx->quantum_state->num_qubits;
-            int qubit3 = (gate_selector >> 16) % ctx->quantum_state->num_qubits;
-            
-            // Apply 4 different gates for stronger mixing
-            gate_hadamard(ctx->quantum_state, qubit1);
-            gate_pauli_z(ctx->quantum_state, qubit2);
-            if (qubit1 != qubit3) {
-                gate_cnot(ctx->quantum_state, qubit1, qubit3);
-            }
-            gate_pauli_x(ctx->quantum_state, (qubit2 + 1) % ctx->quantum_state->num_qubits);
-        }
-        
-        // PERFORMANCE CRITICAL: Batch measure all qubits in ONE pass (8x faster!)
-        uint64_t measurement = quantum_measure_all_fast(
-            ctx->quantum_state,
-            &ctx->entropy_ctx
-        );
-        ctx->stats.quantum_measurements += ctx->config.num_qubits;
-
-        // A basis-state index contains only num_qubits bits (8 by default).
-        // Condition it with an independent full-width entropy word before
-        // serializing it, so repeated Born outcomes do not become repeated
-        // 8-byte output patterns.
-        uint64_t conditioner;
-        if (quantum_entropy_get_uint64(&ctx->entropy_ctx, &conditioner) != 0) {
-            return -1;
-        }
-        measurement ^= conditioner;
-        
-        measurements_since_evolution++;
-        
-        // Extract bytes from measurement
-        size_t bytes_to_copy = sizeof(measurement);
-        if (bytes_to_copy > size - bytes_generated) {
-            bytes_to_copy = size - bytes_generated;
-        }
-        
-        memcpy(buffer + bytes_generated, &measurement, bytes_to_copy);
-        bytes_generated += bytes_to_copy;
-    }
-    
-    return 0;
-}
-
-/**
- * @brief Extract randomness using Grover-enhanced quantum state (OPTIMIZED)
- *
- * KEY INSIGHT: Running full Grover search per byte is wasteful!
- * Instead: Use Grover to prepare interesting quantum distributions,
- * then measure multiple times from that distribution.
- *
- * This gives quantum enhancement WITHOUT the overhead of repeated searches.
- */
-static int extract_grover_entropy(
-    qrng_v3_ctx_t *ctx,
-    uint8_t *buffer,
-    size_t size
-) {
-    if (!ctx || !buffer || size == 0) return -1;
-    
-    size_t bytes_generated = 0;
-    
-    // OPTIMIZATION: Batch extraction - run Grover ONCE, measure MANY times
-    const size_t measurements_per_grover = 128;  // 128 measurements per Grover prep
-    size_t measurements_extracted = measurements_per_grover;  // Force first Grover
-    
-    while (bytes_generated < size) {
-        // Run Grover preparation only when batch depleted
-        if (measurements_extracted >= measurements_per_grover) {
-            // Use Grover to prepare quantum state with interesting distribution
-            // Pick random target to create non-uniform but quantum-evolved state
-            uint64_t random_target;
-            if (quantum_entropy_get_uint64(&ctx->entropy_ctx, &random_target) != 0) {
-                return -1;
-            }
-            random_target = random_target % ctx->quantum_state->state_dim;
-            
-            // Run simplified Grover preparation (fewer iterations for speed)
-            quantum_state_reset(ctx->quantum_state);
-            for (size_t q = 0; q < ctx->config.num_qubits; q++) {
-                gate_hadamard(ctx->quantum_state, q);
-            }
-            
-            // Apply just 2-3 Grover iterations (enough for enhancement, not full search)
-            for (int iter = 0; iter < 2; iter++) {
-                grover_oracle(ctx->quantum_state, random_target);
-                grover_diffusion(ctx->quantum_state);
-            }
-            
-            ctx->stats.grover_searches++;
-            measurements_extracted = 0;
-        }
-        
-        // PERFORMANCE CRITICAL: Batch measure all qubits in ONE pass (8x faster!)
-        uint64_t measurement = quantum_measure_all_fast(
-            ctx->quantum_state,
-            &ctx->entropy_ctx
-        );
-        ctx->stats.quantum_measurements += ctx->config.num_qubits;
-
-        // quantum_measure_all_fast returns a basis index, not 64 bits of
-        // entropy. Apply the same full-width conditioning as direct mode.
-        uint64_t conditioner;
-        if (quantum_entropy_get_uint64(&ctx->entropy_ctx, &conditioner) != 0) {
-            return -1;
-        }
-        measurement ^= conditioner;
-        
-        measurements_extracted++;
-        
-        // Extract bytes
-        size_t bytes_to_copy = sizeof(measurement);
-        if (bytes_to_copy > size - bytes_generated) {
-            bytes_to_copy = size - bytes_generated;
-        }
-        
-        memcpy(buffer + bytes_generated, &measurement, bytes_to_copy);
-        bytes_generated += bytes_to_copy;
-    }
-    
-    return 0;
-}
-
-// ============================================================================
-// INITIALIZATION & CLEANUP
-// ============================================================================
-
-qrng_v3_error_t qrng_v3_init(qrng_v3_ctx_t **ctx_out) {
-    qrng_v3_config_t config;
-    qrng_v3_get_default_config(&config);
-    return qrng_v3_init_with_config(ctx_out, &config);
-}
-
-qrng_v3_error_t qrng_v3_init_with_config(
-    qrng_v3_ctx_t **ctx_out,
-    const qrng_v3_config_t *config
-) {
-    VALIDATE_NOT_NULL(ctx_out, QRNG_V3_ERROR_NULL_CONTEXT);
-    VALIDATE_NOT_NULL(config, QRNG_V3_ERROR_INVALID_PARAM);
-    
-    // Validate configuration
-    if (config->num_qubits < 2 || config->num_qubits > MAX_QUBITS) {
-        return QRNG_V3_ERROR_INVALID_PARAM;
-    }
-    
-    // Allocate context
-    qrng_v3_ctx_t *ctx = calloc(1, sizeof(qrng_v3_ctx_t));
-    if (!ctx) return QRNG_V3_ERROR_OUT_OF_MEMORY;
-    
-    // Copy configuration
-    memcpy(&ctx->config, config, sizeof(qrng_v3_config_t));
-    
-    // LAYER 1: Initialize hardware entropy pool (base layer)
-    entropy_pool_config_t pool_config = {
-        .pool_size = config->entropy_pool_size,
-        .refill_threshold = config->entropy_pool_size / 4,
-        .chunk_size = 4096,
-        .enable_background_thread = config->enable_background_entropy,
-        .min_entropy = 4.0
-    };
-    
-    int pool_err = entropy_pool_init_with_config(&ctx->entropy_pool, &pool_config);
-    if (pool_err != 0) {
-        free(ctx);
-        return QRNG_V3_ERROR_ENTROPY_FAILURE;
-    }
-    
-    // LAYER 2: Initialize quantum simulation engine
-    ctx->quantum_state = calloc(1, sizeof(quantum_state_t));
-    if (!ctx->quantum_state) {
-        entropy_pool_free(ctx->entropy_pool);
-        free(ctx);
-        return QRNG_V3_ERROR_OUT_OF_MEMORY;
-    }
-    
-    qs_error_t qs_err = quantum_state_init(ctx->quantum_state, config->num_qubits);
-    if (qs_err != QS_SUCCESS) {
-        free(ctx->quantum_state);
-        entropy_pool_free(ctx->entropy_pool);
-        free(ctx);
-        return QRNG_V3_ERROR_QUANTUM_INIT;
-    }
-    
-    // Connect entropy pool to quantum measurements (resolves circular dependency!)
-    quantum_entropy_init(
-        &ctx->entropy_ctx,
-        entropy_pool_callback,
-        ctx->entropy_pool
-    );
-    
-    // LAYER 3: Initialize output buffer (larger = better performance)
-    ctx->output_buffer_size = config->output_buffer_size > 0 ?
-                               config->output_buffer_size : 65536;  // Default 64KB
-    ctx->output_buffer = calloc(1, ctx->output_buffer_size);
-    if (!ctx->output_buffer) {
-        quantum_state_free(ctx->quantum_state);
-        free(ctx->quantum_state);
-        entropy_pool_free(ctx->entropy_pool);
-        free(ctx);
-        return QRNG_V3_ERROR_OUT_OF_MEMORY;
-    }
-    ctx->buffer_pos = ctx->output_buffer_size;  // Force initial fill
-    
-    // Initialize Bell test monitoring
-    if (config->enable_bell_monitoring) {
-        ctx->bell_monitor = calloc(1, sizeof(bell_test_monitor_t));
-        if (ctx->bell_monitor) {
-            bell_monitor_init(ctx->bell_monitor, 100);  // Track last 100 tests
-        }
-    }
-    
-    // Initialize Grover cache
-    if (config->enable_grover_cache && config->grover_cache_size > 0) {
-        ctx->grover_cache = calloc(config->grover_cache_size, sizeof(uint64_t));
-    }
-    
-    // Initialize performance monitoring
-    if (config->enable_performance_monitoring) {
-        perf_monitor_init(&ctx->perf_monitor);
-    }
-    
-    ctx->initialized = 1;
-    *ctx_out = ctx;
-    
-    return QRNG_V3_SUCCESS;
-}
-
-void qrng_v3_free(qrng_v3_ctx_t *ctx) {
-    if (!ctx) return;
-    
-    // Free quantum state
-    if (ctx->quantum_state) {
-        quantum_state_free(ctx->quantum_state);
-        free(ctx->quantum_state);
-    }
-    
-    // Free entropy pool
-    if (ctx->entropy_pool) {
-        entropy_pool_free(ctx->entropy_pool);
-    }
-    
-    // Free Bell monitor
-    if (ctx->bell_monitor) {
-        bell_monitor_free(ctx->bell_monitor);
-        free(ctx->bell_monitor);
-    }
-    
-    // Free output buffer
-    if (ctx->output_buffer) {
-        secure_memzero(ctx->output_buffer, ctx->output_buffer_size);
-        free(ctx->output_buffer);
-    }
-    
-    // Free Grover cache
-    if (ctx->grover_cache) {
-        secure_memzero(ctx->grover_cache, ctx->config.grover_cache_size * sizeof(uint64_t));
-        free(ctx->grover_cache);
-    }
-    
-    // Free performance monitor
-    if (ctx->perf_monitor) {
-        perf_monitor_free(ctx->perf_monitor);
-    }
-    
-    // Zero context
-    secure_memzero(ctx, sizeof(*ctx));
-    free(ctx);
-}
-
-// ============================================================================
-// RANDOM NUMBER GENERATION
-// ============================================================================
-
-qrng_v3_error_t qrng_v3_bytes(
-    qrng_v3_ctx_t *ctx,
-    uint8_t *buffer,
-    size_t size
-) {
-    VALIDATE_NOT_NULL(ctx, QRNG_V3_ERROR_NULL_CONTEXT);
-    VALIDATE_BUFFER(buffer, size, QRNG_V3_ERROR_NULL_BUFFER);
-    
-    if (!ctx->initialized) {
-        return QRNG_V3_ERROR_NOT_INITIALIZED;
-    }
-    
-    // Performance monitoring
-    if (ctx->perf_monitor) {
-        perf_monitor_start_operation(ctx->perf_monitor, PERF_OP_OUTPUT_GENERATION);
-    }
-    
-    size_t bytes_copied = 0;
-    
-    while (bytes_copied < size) {
-        // Refill buffer if needed
-        if (ctx->buffer_pos >= ctx->output_buffer_size) {
-            // Generate fresh quantum entropy
-            int err;
-            
-            switch (ctx->config.mode) {
-                case QRNG_V3_MODE_DIRECT:
-                    err = extract_quantum_entropy(ctx, ctx->output_buffer, ctx->output_buffer_size);
-                    break;
-                    
-                case QRNG_V3_MODE_GROVER:
-                    // Grover-based entropy extraction
-                    err = extract_grover_entropy(ctx, ctx->output_buffer, ctx->output_buffer_size);
-                    break;
-                    
-                case QRNG_V3_MODE_BELL_VERIFIED:
-                    err = extract_quantum_entropy(ctx, ctx->output_buffer, ctx->output_buffer_size);
-                    break;
-                    
-                default:
-                    return QRNG_V3_ERROR_INVALID_PARAM;
-            }
-            
-            if (err != 0) {
-                if (ctx->perf_monitor) {
-                    perf_monitor_end_operation(ctx->perf_monitor);
-                }
-                return QRNG_V3_ERROR_ENTROPY_FAILURE;
-            }
-            
-            ctx->buffer_pos = 0;
-        }
-        
-        // Copy from buffer
-        size_t copy_size = ctx->output_buffer_size - ctx->buffer_pos;
-        if (copy_size > size - bytes_copied) {
-            copy_size = size - bytes_copied;
-        }
-        
-        memcpy(buffer + bytes_copied, ctx->output_buffer + ctx->buffer_pos, copy_size);
-        ctx->buffer_pos += copy_size;
-        bytes_copied += copy_size;
-    }
-    
-    // Update statistics
-    ctx->stats.bytes_generated += size;
-    ctx->bytes_since_bell_test += size;
-    
-    // Bell test monitoring
-    if (ctx->config.enable_bell_monitoring && 
-        ctx->config.bell_test_interval > 0 &&
-        ctx->bytes_since_bell_test >= ctx->config.bell_test_interval) {
-        
-        // Use enough measurements that the CHSH estimate is statistically
-        // tight (SE ~ 2/sqrt(N)); a small N made monitored generation
-        // spuriously trip the min-CHSH check on an unlucky draw.
-        bell_test_result_t result = qrng_v3_verify_quantum(ctx, 4000);
-
-        if (ctx->bell_monitor) {
-            bell_monitor_add_result(ctx->bell_monitor, &result);
-        }
-        
-        ctx->stats.bell_tests_performed++;
-        if (bell_test_confirms_quantum(&result)) {
-            ctx->stats.bell_tests_passed++;
-        }
-        
-        // Check if quantum behavior is maintained
-        if (result.chsh_value < ctx->config.min_acceptable_chsh) {
-            // Quantum behavior degraded - this shouldn't happen in simulation
-            // but good to check
-            return QRNG_V3_ERROR_BELL_TEST_FAILED;
-        }
-        
-        ctx->bytes_since_bell_test = 0;
-    }
-    
-    if (ctx->perf_monitor) {
-        perf_monitor_end_operation(ctx->perf_monitor);
-        perf_monitor_record_bytes(ctx->perf_monitor, size);
-    }
-    
-    return QRNG_V3_SUCCESS;
-}
-
-qrng_v3_error_t qrng_v3_uint64(qrng_v3_ctx_t *ctx, uint64_t *value) {
-    VALIDATE_NOT_NULL(ctx, QRNG_V3_ERROR_NULL_CONTEXT);
-    VALIDATE_NOT_NULL(value, QRNG_V3_ERROR_NULL_BUFFER);
-    
-    return qrng_v3_bytes(ctx, (uint8_t*)value, sizeof(*value));
-}
-
-qrng_v3_error_t qrng_v3_double(qrng_v3_ctx_t *ctx, double *value) {
-    VALIDATE_NOT_NULL(ctx, QRNG_V3_ERROR_NULL_CONTEXT);
-    VALIDATE_NOT_NULL(value, QRNG_V3_ERROR_NULL_BUFFER);
-    
-    uint64_t random_bits;
-    qrng_v3_error_t err = qrng_v3_uint64(ctx, &random_bits);
-    if (err != QRNG_V3_SUCCESS) return err;
-    
-    // Convert to double in [0, 1) using 53 bits of precision
-    *value = (double)(random_bits >> 11) * 0x1.0p-53;
-    
-    return QRNG_V3_SUCCESS;
-}
-
-qrng_v3_error_t qrng_v3_range(
-    qrng_v3_ctx_t *ctx,
-    uint64_t min,
-    uint64_t max,
-    uint64_t *value
-) {
-    VALIDATE_NOT_NULL(ctx, QRNG_V3_ERROR_NULL_CONTEXT);
-    VALIDATE_NOT_NULL(value, QRNG_V3_ERROR_NULL_BUFFER);
-    
-    if (min > max) return QRNG_V3_ERROR_INVALID_PARAM;
-    if (min == max) {
-        *value = min;
-        return QRNG_V3_SUCCESS;
-    }
-    
-    uint64_t range = max - min + 1;
-    if (range == 0) {  // Overflow case
-        qrng_v3_uint64(ctx, value);
-        return QRNG_V3_SUCCESS;
-    }
-    
-    // Unbiased sampling
-    uint64_t threshold = (UINT64_MAX - range + 1) % range;
-    uint64_t r;
-    
-    do {
-        qrng_v3_error_t err = qrng_v3_uint64(ctx, &r);
-        if (err != QRNG_V3_SUCCESS) return err;
-    } while (r < threshold);
-    
-    *value = min + (r % range);
-    return QRNG_V3_SUCCESS;
-}
-
-// ============================================================================
-// GROVER-ENHANCED SAMPLING
-// ============================================================================
-
-qrng_v3_error_t qrng_v3_grover_sample(
-    qrng_v3_ctx_t *ctx,
-    uint64_t *value
-) {
-    VALIDATE_NOT_NULL(ctx, QRNG_V3_ERROR_NULL_CONTEXT);
-    VALIDATE_NOT_NULL(value, QRNG_V3_ERROR_NULL_BUFFER);
-    
-    if (!ctx->initialized) {
-        return QRNG_V3_ERROR_NOT_INITIALIZED;
-    }
-    
-    // Use Grover's algorithm for quantum sampling
-    *value = grover_random_sample(
-        ctx->quantum_state,
-        ctx->config.num_qubits,
-        &ctx->entropy_ctx
-    );
-    
-    ctx->stats.grover_searches++;
-    
-    return QRNG_V3_SUCCESS;
-}
-
-qrng_v3_error_t qrng_v3_grover_sample_distribution(
-    qrng_v3_ctx_t *ctx,
-    double (*target_distribution)(uint64_t),
-    uint64_t *value
-) {
-    VALIDATE_NOT_NULL(ctx, QRNG_V3_ERROR_NULL_CONTEXT);
-    VALIDATE_NOT_NULL(target_distribution, QRNG_V3_ERROR_NULL_BUFFER);
-    VALIDATE_NOT_NULL(value, QRNG_V3_ERROR_NULL_BUFFER);
-    
-    /**
-     * FULL PRODUCTION AMPLITUDE AMPLIFICATION
-     *
-     * This implements quantum amplitude amplification to sample from
-     * arbitrary probability distributions. This is a generalization of
-     * Grover's algorithm for non-uniform distributions.
-     *
-     * Algorithm (Brassard et al., "Quantum Amplitude Amplification"):
-     * 1. Prepare state with amplitudes ∝ √P(x) where P is target distribution
-     * 2. Apply amplitude amplification operator iteratively
-     * 3. Measure to sample from P(x)
-     *
-     * Quantum advantage: Fewer iterations than rejection sampling for
-     * peaked distributions.
-     */
-    
-    quantum_state_t *state = ctx->quantum_state;
-    uint64_t state_dim = state->state_dim;
-    
-    // Step 1: Initialize to uniform superposition
-    quantum_state_reset(state);
-    for (size_t q = 0; q < ctx->config.num_qubits; q++) {
-        gate_hadamard(state, q);
-    }
-    
-    // Step 2: Encode target distribution as amplitude modulation
-    // We set amplitude[i] ∝ sqrt(target_distribution(i))
-    double normalization = 0.0;
-    
-    for (uint64_t i = 0; i < state_dim; i++) {
-        double target_prob = target_distribution(i);
-        
-        // Ensure valid probability
-        if (target_prob < 0.0) target_prob = 0.0;
-        if (target_prob > 1.0) target_prob = 1.0;
-        
-        // Born rule: |amplitude|^2 = probability
-        // Therefore: amplitude = sqrt(probability) * phase_factor
-        double target_amplitude = sqrt(target_prob);
-        
-        // Preserve existing phase from uniform superposition
-        complex_t current = state->amplitudes[i];
-        double current_mag = cabs(current);
-        
-        if (current_mag > 1e-15) {
-            // Scale amplitude while preserving phase
-            complex_t phase_factor = current / current_mag;
-            state->amplitudes[i] = target_amplitude * phase_factor;
-        } else {
-            // No phase information, use real amplitude
-            state->amplitudes[i] = target_amplitude + 0.0 * I;
-        }
-        
-        normalization += target_amplitude * target_amplitude;
-    }
-    
-    // Step 3: Renormalize state (required for valid quantum state)
-    if (normalization > 1e-15) {
-        double norm = sqrt(normalization);
-        for (uint64_t i = 0; i < state_dim; i++) {
-            state->amplitudes[i] /= norm;
-        }
-    }
-    
-    // Step 4: Apply Grover diffusion to amplify target amplitudes
-    // Number of iterations for amplitude amplification:
-    // optimal = π/4 * sqrt(N/M) where N=state_dim, M=effective good states
-    
-    // Estimate M from distribution (effective number of states with high probability)
-    double effective_good_states = 0.0;
-    double max_prob = 0.0;
-    
-    for (uint64_t i = 0; i < state_dim; i++) {
-        double prob = target_distribution(i);
-        if (prob > max_prob) max_prob = prob;
-    }
-    
-    for (uint64_t i = 0; i < state_dim; i++) {
-        if (target_distribution(i) > max_prob * 0.1) {  // 10% threshold
-            effective_good_states += 1.0;
-        }
-    }
-    
-    if (effective_good_states < 1.0) effective_good_states = 1.0;
-    
-    // Calculate optimal iterations
-    size_t num_iterations = (size_t)(QC_PI_4 * sqrt((double)state_dim / effective_good_states));
-    
-    // Practical limits
-    if (num_iterations < 1) num_iterations = 1;
-    if (num_iterations > 50) num_iterations = 50;
-    
-    // Apply Grover diffusion operators
-    for (size_t iter = 0; iter < num_iterations; iter++) {
-        // Grover diffusion: inverts amplitudes about their average
-        // This amplifies amplitudes above average, suppresses below average
-        qs_error_t err = grover_diffusion(state);
-        if (err != QS_SUCCESS) {
-            return QRNG_V3_ERROR_QUANTUM_INIT;
-        }
-    }
-    
-    // Step 5: Measure to sample from amplified distribution
-    // Use hardware entropy for measurement sampling
-    double random;
-    if (quantum_entropy_get_double(&ctx->entropy_ctx, &random) != 0) {
-        return QRNG_V3_ERROR_ENTROPY_FAILURE;
-    }
-    
-    // Sample from final probability distribution
-    double cumulative = 0.0;
-    for (uint64_t i = 0; i < state_dim; i++) {
-        double prob = quantum_state_get_probability(state, i);
-        cumulative += prob;
-        
-        if (random < cumulative) {
-            *value = i;
-            ctx->stats.grover_searches++;
-            return QRNG_V3_SUCCESS;
-        }
-    }
-    
-    // Fallback (numerical precision edge case)
-    *value = state_dim - 1;
-    ctx->stats.grover_searches++;
-    
-    return QRNG_V3_SUCCESS;
-}
-
-qrng_v3_error_t qrng_v3_grover_multi_target(
-    qrng_v3_ctx_t *ctx,
-    const uint64_t *targets,
-    size_t num_targets,
-    size_t *found_index,
-    uint64_t *value
-) {
-    VALIDATE_NOT_NULL(ctx, QRNG_V3_ERROR_NULL_CONTEXT);
-    VALIDATE_NOT_NULL(targets, QRNG_V3_ERROR_NULL_BUFFER);
-    VALIDATE_NOT_NULL(found_index, QRNG_V3_ERROR_NULL_BUFFER);
-    VALIDATE_NOT_NULL(value, QRNG_V3_ERROR_NULL_BUFFER);
-    
-    if (num_targets == 0) return QRNG_V3_ERROR_INVALID_PARAM;
-    
-    /**
-     * MULTI-TARGET GROVER SEARCH (Production Implementation)
-     *
-     * Uses Grover's algorithm to search for multiple targets simultaneously.
-     * Oracle marks ALL targets, diffusion amplifies them collectively.
-     *
-     * Speedup: √(N/M) where N=search space, M=number of targets
-     */
-    
-    quantum_state_t *state = ctx->quantum_state;
-    uint64_t state_dim = state->state_dim;
-    
-    // Validate all targets are within state space
-    for (size_t i = 0; i < num_targets; i++) {
-        if (targets[i] >= state_dim) {
-            return QRNG_V3_ERROR_INVALID_PARAM;
-        }
-    }
-    
-    // Step 1: Initialize to uniform superposition
-    quantum_state_reset(state);
-    for (size_t q = 0; q < ctx->config.num_qubits; q++) {
-        gate_hadamard(state, q);
-    }
-    
-    // Step 2: Calculate optimal iterations for multi-target search
-    // optimal_iterations = π/4 * sqrt(N/M)
-    double ratio = (double)state_dim / (double)num_targets;
-    size_t iterations = (size_t)(QC_PI_4 * sqrt(ratio));
-    
-    if (iterations < 1) iterations = 1;
-    if (iterations > 100) iterations = 100;  // Safety limit
-    
-    // Step 3: Apply Grover iterations
-    for (size_t iter = 0; iter < iterations; iter++) {
-        // Oracle: Mark all target states with phase flip
-        for (size_t t = 0; t < num_targets; t++) {
-            qs_error_t err = grover_oracle(state, targets[t]);
-            if (err != QS_SUCCESS) {
-                return QRNG_V3_ERROR_QUANTUM_INIT;
-            }
-        }
-        
-        // Diffusion: Amplitude amplification
-        qs_error_t err = grover_diffusion(state);
-        if (err != QS_SUCCESS) {
-            return QRNG_V3_ERROR_QUANTUM_INIT;
-        }
-    }
-    
-    // Step 4: Measure to find one of the targets
-    double random;
-    if (quantum_entropy_get_double(&ctx->entropy_ctx, &random) != 0) {
-        return QRNG_V3_ERROR_ENTROPY_FAILURE;
-    }
-    
-    // Sample from amplified probability distribution
-    double cumulative = 0.0;
-    uint64_t measured_state = 0;
-    
-    for (uint64_t i = 0; i < state_dim; i++) {
-        double prob = quantum_state_get_probability(state, i);
-        cumulative += prob;
-        
-        if (random < cumulative) {
-            measured_state = i;
-            break;
-        }
-    }
-    
-    // Step 5: Determine which target was found
-    for (size_t t = 0; t < num_targets; t++) {
-        if (measured_state == targets[t]) {
-            *found_index = t;
-            *value = targets[t];
-            ctx->stats.grover_searches++;
-            return QRNG_V3_SUCCESS;
-        }
-    }
-    
-    // If we didn't find an exact target (rare but possible), return closest
-    // Find target with highest probability
-    size_t best_target = 0;
-    double best_prob = 0.0;
-    
-    for (size_t t = 0; t < num_targets; t++) {
-        double prob = quantum_state_get_probability(state, targets[t]);
-        if (prob > best_prob) {
-            best_prob = prob;
-            best_target = t;
-        }
-    }
-    
-    *found_index = best_target;
-    *value = targets[best_target];
-    ctx->stats.grover_searches++;
-    
-    return QRNG_V3_SUCCESS;
-}
-
-// ============================================================================
-// QUANTUM VERIFICATION
-// ============================================================================
-
-bell_test_result_t qrng_v3_verify_quantum(
-    qrng_v3_ctx_t *ctx,
-    size_t num_measurements
-) {
-    bell_test_result_t result = {0};
-    
-    if (!ctx || !ctx->initialized || !ctx->quantum_state) {
-        return result;
-    }
-    
-    // Run Bell test on current quantum state
-    result = bell_test_chsh(
-        ctx->quantum_state,
-        0,  // Qubit A
-        1,  // Qubit B
-        num_measurements,
-        NULL,  // Use optimal settings
-        &ctx->entropy_ctx
-    );
-    
-    // Update statistics
-    if (result.chsh_value > ctx->stats.max_chsh) {
-        ctx->stats.max_chsh = result.chsh_value;
-    }
-    if (ctx->stats.min_chsh == 0.0 || result.chsh_value < ctx->stats.min_chsh) {
-        ctx->stats.min_chsh = result.chsh_value;
-    }
-    
-    // Update running average
-    double total = ctx->stats.average_chsh * ctx->stats.bell_tests_performed;
-    total += result.chsh_value;
-    ctx->stats.bell_tests_performed++;
-    ctx->stats.average_chsh = total / ctx->stats.bell_tests_performed;
-    
-    return result;
-}
-
-double qrng_v3_get_entanglement_entropy(const qrng_v3_ctx_t *ctx) {
-    if (!ctx || !ctx->quantum_state) return 0.0;
-    
-    // Calculate entanglement between first half and second half of qubits
-    size_t half = ctx->config.num_qubits / 2;
-    if (half == 0) return 0.0;
-    
-    int *subsystem_a = calloc(half, sizeof(int));
-    if (!subsystem_a) return 0.0;
-    
-    for (size_t i = 0; i < half; i++) {
-        subsystem_a[i] = i;
-    }
-    
-    double entropy = quantum_state_entanglement_entropy(
-        ctx->quantum_state,
-        subsystem_a,
-        half
-    );
-    
-    free(subsystem_a);
-    
-    // Numerical safety: entanglement entropy must be non-negative
-    // Negative values from numerical precision issues should be clamped
-    if (entropy < 0.0 || isnan(entropy)) {
-        entropy = 0.0;  // Physically valid lower bound
-    }
-    
-    return entropy;
-}
-
-int qrng_v3_is_quantum_verified(const qrng_v3_ctx_t *ctx) {
-    if (!ctx || !ctx->bell_monitor) return 0;
-    
-    // Check if recent Bell tests confirm quantum behavior
-    if (ctx->stats.bell_tests_performed == 0) return 0;
-    
-    // At least 80% of tests should pass
-    double pass_rate = (double)ctx->stats.bell_tests_passed / ctx->stats.bell_tests_performed;
-    if (pass_rate < 0.8) return 0;
-    
-    // Average CHSH should be above classical bound
-    if (ctx->stats.average_chsh <= 2.0) return 0;
-    
-    return 1;
-}
-
-// ============================================================================
-// MODE CONTROL
-// ============================================================================
-
-qrng_v3_error_t qrng_v3_set_mode(qrng_v3_ctx_t *ctx, qrng_v3_mode_t mode) {
-    VALIDATE_NOT_NULL(ctx, QRNG_V3_ERROR_NULL_CONTEXT);
-    
-    ctx->config.mode = mode;
-    return QRNG_V3_SUCCESS;
-}
-
-qrng_v3_mode_t qrng_v3_get_mode(const qrng_v3_ctx_t *ctx) {
-    if (!ctx) return QRNG_V3_MODE_DIRECT;
-    return ctx->config.mode;
-}
-
-// ============================================================================
-// STATISTICS & MONITORING
-// ============================================================================
-
-qrng_v3_error_t qrng_v3_get_stats(
-    const qrng_v3_ctx_t *ctx,
-    qrng_v3_stats_t *stats
-) {
-    VALIDATE_NOT_NULL(ctx, QRNG_V3_ERROR_NULL_CONTEXT);
-    VALIDATE_NOT_NULL(stats, QRNG_V3_ERROR_NULL_BUFFER);
-    
-    memcpy(stats, &ctx->stats, sizeof(*stats));
-    
-    // Calculate current entanglement
-    stats->current_entanglement_entropy = qrng_v3_get_entanglement_entropy(ctx);
-    
-    return QRNG_V3_SUCCESS;
-}
-
-void qrng_v3_print_stats(const qrng_v3_ctx_t *ctx) {
-    if (!ctx) return;
-    
-    printf("\n");
-    printf("╔═══════════════════════════════════════════════════════════╗\n");
-    printf("║         QUANTUM RNG v3.0 STATISTICS                       ║\n");
-    printf("╠═══════════════════════════════════════════════════════════╣\n");
-    printf("║                                                           ║\n");
-    printf("║  Generation:                                              ║\n");
-    printf("║    Bytes generated:     %10llu                         ║\n",
-           (unsigned long long)ctx->stats.bytes_generated);
-    printf("║    Quantum measurements:%10llu                         ║\n",
-           (unsigned long long)ctx->stats.quantum_measurements);
-    printf("║    Grover searches:     %10llu                         ║\n",
-           (unsigned long long)ctx->stats.grover_searches);
-    printf("║                                                           ║\n");
-    printf("║  Quantum Verification:                                    ║\n");
-    printf("║    Bell tests performed:%10llu                         ║\n",
-           (unsigned long long)ctx->stats.bell_tests_performed);
-    printf("║    Bell tests passed:   %10llu                         ║\n",
-           (unsigned long long)ctx->stats.bell_tests_passed);
-    
-    if (ctx->stats.bell_tests_performed > 0) {
-        printf("║    Average CHSH:        %10.4f                         ║\n",
-               ctx->stats.average_chsh);
-        printf("║    Min CHSH:            %10.4f                         ║\n",
-               ctx->stats.min_chsh);
-        printf("║    Max CHSH:            %10.4f                         ║\n",
-               ctx->stats.max_chsh);
-        
-        double pass_rate = 100.0 * ctx->stats.bell_tests_passed / ctx->stats.bell_tests_performed;
-        printf("║    Pass rate:           %9.1f%%                        ║\n", pass_rate);
-    }
-    
-    printf("║                                                           ║\n");
-    printf("║  Quantum Properties:                                      ║\n");
-    
-    double ent_entropy = qrng_v3_get_entanglement_entropy(ctx);
-    printf("║    Entanglement entropy:%10.4f bits                    ║\n", ent_entropy);
-    printf("║    State entropy:       %10.4f bits                    ║\n",
-           quantum_state_entropy(ctx->quantum_state));
-    printf("║    Purity:              %10.4f                         ║\n",
-           quantum_state_purity(ctx->quantum_state));
-    
-    printf("║                                                           ║\n");
-    printf("║  Mode: %-50s ║\n", qrng_v3_mode_string(ctx->config.mode));
-    printf("║  Quantum Verified: %-42s ║\n",
-           qrng_v3_is_quantum_verified(ctx) ? "✓ YES" : "✗ NO");
-    
-    printf("╚═══════════════════════════════════════════════════════════╝\n");
-    printf("\n");
-    
-    // Print entropy pool stats
-    if (ctx->entropy_pool) {
-        entropy_pool_print_stats(ctx->entropy_pool);
-    }
-    
-    // Print performance stats
-    if (ctx->perf_monitor) {
-        perf_monitor_print_stats(ctx->perf_monitor);
-    }
-}
-
-const bell_test_monitor_t* qrng_v3_get_bell_history(const qrng_v3_ctx_t *ctx) {
-    if (!ctx) return NULL;
-    return ctx->bell_monitor;
-}
-
-// ============================================================================
-// UTILITY FUNCTIONS
-// ============================================================================
-
-const char* qrng_v3_version(void) {
-    static char version[32];
-    snprintf(version, sizeof(version), "%d.%d.%d",
-             QUANTUM_RNG_V3_VERSION_MAJOR,
-             QUANTUM_RNG_V3_VERSION_MINOR,
-             QUANTUM_RNG_V3_VERSION_PATCH);
-    return version;
-}
-
-const char* qrng_v3_error_string(qrng_v3_error_t error) {
-    switch (error) {
-        case QRNG_V3_SUCCESS:
-            return "Success";
-        case QRNG_V3_ERROR_NULL_CONTEXT:
-            return "NULL context provided";
-        case QRNG_V3_ERROR_NULL_BUFFER:
-            return "NULL buffer provided";
-        case QRNG_V3_ERROR_INVALID_PARAM:
-            return "Invalid parameter";
-        case QRNG_V3_ERROR_ENTROPY_FAILURE:
-            return "Entropy generation failed";
-        case QRNG_V3_ERROR_QUANTUM_INIT:
-            return "Quantum state initialization failed";
-        case QRNG_V3_ERROR_BELL_TEST_FAILED:
-            return "Bell test verification failed - quantum behavior not confirmed";
-        case QRNG_V3_ERROR_OUT_OF_MEMORY:
-            return "Out of memory";
-        case QRNG_V3_ERROR_NOT_INITIALIZED:
-            return "Context not initialized";
-        default:
-            return "Unknown error";
-    }
-}
-
-const char* qrng_v3_mode_string(qrng_v3_mode_t mode) {
-    switch (mode) {
-        case QRNG_V3_MODE_DIRECT:
-            return "DIRECT (quantum measurements)";
-        case QRNG_V3_MODE_GROVER:
-            return "GROVER (quantum search sampling)";
-        case QRNG_V3_MODE_BELL_VERIFIED:
-            return "BELL_VERIFIED (continuous verification)";
-        default:
-            return "Unknown mode";
-    }
-}
-
-// ============================================================================
-// BACKWARD COMPATIBILITY
-// ============================================================================
-
-qrng_v3_error_t qrng_v3_init_from_seed(
-    qrng_v3_ctx_t **ctx,
-    const uint8_t *seed,
-    size_t seed_len
-) {
-    // Initialize with default config
-    qrng_v3_error_t err = qrng_v3_init(ctx);
-    if (err != QRNG_V3_SUCCESS) return err;
-    
-    // Use seed to initialize quantum state if provided
-    if (seed && seed_len > 0) {
-        // Apply seed-based gates to quantum state
-        for (size_t i = 0; i < seed_len && i < (*ctx)->config.num_qubits; i++) {
-            double angle = (seed[i] / 255.0) * QC_2PI;
-            gate_ry((*ctx)->quantum_state, i % (*ctx)->config.num_qubits, angle);
-        }
-    }
-    
-    return QRNG_V3_SUCCESS;
-}
